test: add test for guest product controller, get all products and get product by id
